pyproject.toml
[project]
name = "mlops-ref"
version = "0.1.0"
description = "Reference mlops pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform==1.128.0",
    "kfp==2.14.6",
    "google-cloud-pipeline-components>=2.22.0",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "google-cloud-storage>=3.6.0",
    "google-cloud-bigquery>=3.38.0",
    "db-dtypes>=1.4.4",
    "pyarrow>=22.0.0",
    "numpy>=2.3.5",
    "scipy>=1.16.3",
    "joblib>=1.5.2",
    "dotenv"
]
[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.8",
    "nest-asyncio>=1.6.0",
    "types-requests"
]

[project.optional-dependencies]


jupyter = [
    "jupyter~=1.0.0",
]
lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.0.20240914",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["app", "frontend"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "src"
asyncio_default_fixture_loop_scope = "function"
